refactored package organisation
