+++ core/src/com/batherphilippa/saunscapades/domain/sprite/Character.java
@@ -1,4 +1,4 @@
-package com.batherphilippa.saunscapades.domain;
+package com.batherphilippa.saunscapades.domain.sprite;
 
 import com.badlogic.gdx.graphics.g2d.Sprite;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
