+++ core/src/com/batherphilippa/saunscapades/domain/sprite/Shaun.java
@@ -1,4 +1,4 @@
-package com.batherphilippa.saunscapades.domain;
+package com.batherphilippa.saunscapades.domain.sprite;
 
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.physics.box2d.World;
